Move file name manipulation logic to StorageManager.
