+++ functions/src/RequireOnly.ts
@@ -1,0 +1,11 @@
+/*!
+ * sri sri guru gauranga jayatah
+ */
+
+// Inspired by https://medium.com/@KevinBGreene/typescript-modeling-required-fields-with-mapped-types-f7bf17688786
+export type RequireOnly<T, K extends keyof T> = {
+  [X in Exclude<keyof T, K>]?: T[X];
+} &
+  {
+    [P in K]-?: T[P];
+  };
