+++ functions/src/SoundQualityReporting/TasksRepository.ts
@@ -7,8 +7,8 @@ import { DateTime } from 'luxon';
 import { createSchema, morphism } from 'morphism';
 import { AllotmentStatus } from '../Allotment';
 import { DateTimeConverter } from '../DateTimeConverter';
-import { RequireOnly } from '../helpers';
 import { ReportingTask } from '../ReportingTask';
+import { RequireOnly } from '../RequireOnly';
 import { Spreadsheet } from '../Spreadsheet';
 import admin = require('firebase-admin');
 import _ = require('lodash');
