+++ functions/src/StorageManager.ts
@@ -6,12 +6,47 @@
 import { File } from '@google-cloud/storage';
 import * as functions from 'firebase-functions';
 import { URL } from 'url';
-import { extractListFromFilename } from './helpers';
 import admin = require('firebase-admin');
 
 export type BucketName = 'original' | 'edited' | 'te.uploads';
 
+export const taskIdRegex = '^[a-zA-Z]+-\\d+';
+
 export class StorageManager {
+  static extractListFromFilename = (fileName: string): string => {
+    const match = fileName.match(/^\w+(?=-)|Hi(?=\d)/i);
+    if (!match) return null;
+
+    const list = match[0].toUpperCase();
+    return list === 'HI' ? 'ML1' : list;
+  };
+
+  static standardizeFileName = (fileName: string) => {
+    if (fileName.startsWith('ML2-')) return fileName.replace(/^ML2-/, '');
+
+    return fileName
+      .replace(/^Hi/i, 'ML1-')
+      .replace(
+        /^(ML[12]|[a-zA-Z]+)-(\d{1,4})\s*([\w\s]*)(\.\w{3,4})$/i,
+        (
+          match,
+          list: string,
+          serial: string,
+          suffix: string,
+          extension: string,
+          index,
+          original
+        ) =>
+          [
+            list.toUpperCase(),
+            '-',
+            serial.padStart(list === 'ML2' ? 4 : 3, '0'),
+            suffix.toUpperCase().replace(/[-\s]/g, ''),
+            extension,
+          ].join('')
+      );
+  };
+
   static getFullBucketName(bucketName: BucketName) {
     return `${bucketName}.${functions.config().project.domain}`;
   }
@@ -30,7 +65,7 @@ export class StorageManager {
   static getFile(bucketName: BucketName, fileName: string) {
     return new File(
       this.getBucket(bucketName),
-      `${extractListFromFilename(fileName)}/${fileName}`
+      `${this.extractListFromFilename(fileName)}/${fileName}`
     );
   }
