+++ functions/src/modules/Files.functions.ts
@@ -5,7 +5,6 @@
 import functions = require('firebase-functions');
 import express = require('express');
 import { DateTime } from 'luxon';
-import { standardizeFileName } from '../helpers';
 import { StorageManager } from '../StorageManager';
 
 const app = express();
@@ -14,17 +13,19 @@ app.get('/download/:bucket/:fileName', async (req, res) => {
   const file = StorageManager.getFile(
     <any>req.params.bucket,
     req.params.bucket === 'original'
-      ? standardizeFileName(req.params.fileName)
+      ? StorageManager.standardizeFileName(req.params.fileName)
       : req.params.fileName
   );
   if ((await file.exists())[0]) {
-    const url = (await file.getSignedUrl({
-      action: 'read',
-      expires: DateTime.local()
-        .plus({ days: 3 })
-        .toJSDate(),
-      promptSaveAs: req.params.fileName,
-    }))[0];
+    const url = (
+      await file.getSignedUrl({
+        action: 'read',
+        expires: DateTime.local()
+          .plus({ days: 3 })
+          .toJSDate(),
+        promptSaveAs: req.params.fileName,
+      })
+    )[0];
     console.log(`Redirecting ${req.params.fileName} to ${url}`);
     res.redirect(307, url);
   } else {
